Unifying system utils into base package
