+++ module/asset-s3/src/source.ts
@@ -1,8 +1,8 @@
 import * as aws from 'aws-sdk';
-import * as fs from 'fs';
 import { TagSet } from 'aws-sdk/clients/s3';
 import { Readable } from 'stream';
 
+import { SystemUtil } from '@travetto/base';
 import { AssetSource, Asset, AssetMetadata } from '@travetto/asset';
 import { Injectable } from '@travetto/di';
 
@@ -54,7 +54,7 @@ export class S3AssetSource extends AssetSource {
 
   async write(file: Asset, stream: NodeJS.ReadableStream): Promise<Asset> {
     const upload = this.client.upload(this.q(file.path, {
-      Body: fs.createReadStream(file.path),
+      Body: stream,
       ContentType: file.contentType,
       ContentLength: file.size
     })).promise();
@@ -70,12 +70,10 @@ export class S3AssetSource extends AssetSource {
 
   async read(filename: string): Promise<NodeJS.ReadableStream | Readable> {
     const res = await this.client.getObject({ Bucket: this.config.bucket, Key: filename }).promise();
-    if (res.Body instanceof Buffer || typeof res.Body === 'string') {
-      const strm = new Readable();
-      strm._read = () => { };
-      strm.push(res.Body);
-      strm.push(null);
-      return strm;
+    if (res.Body instanceof Buffer) {
+      return SystemUtil.toReadable(res.Body);
+    } else if (typeof res.Body === 'string') {
+      return SystemUtil.toReadable(Buffer.from(res.Body, 'utf8'));
     } else if (res.Body && ('pipe' in res.Body)) {
       return res.Body as NodeJS.ReadableStream;
     }
