+++ module/base/index.ts
@@ -6,6 +6,7 @@ export * from './src/stacktrace';
 export * from './src/error';
 export * from './src/shutdown';
 export * from './src/util';
+export * from './src/system-util';
 export * from './src/scan-app';
 export * from './src/scan-fs';
 export * from './src/watch';
